pyproject.toml
[project]
name = "hazmat-classifier"
version = "0.1.0"
description = "Hazmat classifier for Mercado Libre products: ensemble ML (XGBoost + LightGBM + RF) with semantic embeddings and LLM fallback"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "pandas>=2.2",
    "pyarrow>=15.0",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "google-genai>=1.0",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
    "joblib>=1.3",
    "mlflow>=2.10",
    "fpdf2>=2.8.7",
    "sentence-transformers>=5.3.0",
    "lightgbm>=4.6.0",
    "optuna>=4.8.0",
    "datasets>=4.8.4",
    "accelerate>=1.13.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "gradio>=5.0",
    "shap>=0.51.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "nbconvert>=7.17.0",
    "ipykernel>=7.2.0",
]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]
lint = [
    "ruff>=0.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
lint = [
    "ruff>=0.15.8",
]
test = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E402", "E501"]

[tool.ruff.lint.per-file-ignores]
"src/api.py" = ["E402"]
"src/dashboard.py" = ["E402"]
"src/llm_fallback.py" = ["E501"]
